+++ packages/ses/src/lockdown-shim.js
@@ -61,6 +61,13 @@ export function repairIntrinsics(
   // Thus, all present options must agree with first options.
   // Reconstructing `option` here also ensures that it is a well
   // behaved record, with only own data properties.
+  //
+  // The `'overrideTaming'` is not a safety issue. Rather it is a tradeoff
+  // between code compatibility, which is better with the `'moderate'`
+  // setting, and tool compatibility, which is better with the `'min'`
+  // setting. See
+  // https://github.com/Agoric/SES-shim/blob/master/packages/ses/README.md#enabling-override-by-assignment)
+  // for an explanation of when to use which.
   options = { ...firstOptions, ...options };
   const {
     dateTaming = 'safe',
@@ -69,6 +76,7 @@ export function repairIntrinsics(
     regExpTaming = 'safe',
     localeTaming = 'safe',
     consoleTaming = 'safe',
+    overrideTaming = 'moderate',
 
     ...extraOptions
   } = options;
@@ -99,6 +107,7 @@ export function repairIntrinsics(
     regExpTaming,
     localeTaming,
     consoleTaming,
+    overrideTaming,
   };
 
   /**
@@ -168,7 +177,7 @@ export function repairIntrinsics(
 
   function hardenIntrinsics() {
     // Circumvent the override mistake.
-    enablePropertyOverrides(intrinsics);
+    enablePropertyOverrides(intrinsics, overrideTaming);
 
     // Finally register and optionally freeze all the intrinsics. This
     // must be the operation that modifies the intrinsics.